refactor: rename pympdata libs
