+++ libs/pympdata_bulk/bulk_scheme_condensation.py
